+++ lib/src/material/material_desktop_controls.dart
@@ -1,16 +1,16 @@
 import 'dart:async';
 
-import 'package:chewie/src/animated_play_pause.dart';
-import 'package:chewie/src/center_play_button.dart';
-import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/chewie_progress_colors.dart';
-import 'package:chewie/src/helpers/utils.dart';
-import 'package:chewie/src/material/material_progress_bar.dart';
-import 'package:chewie/src/material/widgets/options_dialog.dart';
-import 'package:chewie/src/material/widgets/playback_speed_dialog.dart';
-import 'package:chewie/src/models/option_item.dart';
-import 'package:chewie/src/models/subtitle_model.dart';
-import 'package:chewie/src/notifiers/index.dart';
+import 'package:chewie_audio/src/animated_play_pause.dart';
+import 'package:chewie_audio/src/center_play_button.dart';
+import 'package:chewie_audio/src/chewie_player.dart';
+import 'package:chewie_audio/src/chewie_progress_colors.dart';
+import 'package:chewie_audio/src/helpers/utils.dart';
+import 'package:chewie_audio/src/material/material_progress_bar.dart';
+import 'package:chewie_audio/src/material/widgets/options_dialog.dart';
+import 'package:chewie_audio/src/material/widgets/playback_speed_dialog.dart';
+import 'package:chewie_audio/src/models/option_item.dart';
+import 'package:chewie_audio/src/models/subtitle_model.dart';
+import 'package:chewie_audio/src/notifiers/index.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:video_player/video_player.dart';
@@ -48,10 +48,10 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
   final marginSize = 5.0;
 
   late VideoPlayerController controller;
-  ChewieController? _chewieController;
+  ChewieAudioController? _chewieController;
 
   // We know that _chewieController is set in didChangeDependencies
-  ChewieController get chewieController => _chewieController!;
+  ChewieAudioController get chewieController => _chewieController!;
 
   @override
   void initState() {
@@ -88,9 +88,7 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
               if (_displayBufferingIndicator)
                 const Center(
                   child: CircularProgressIndicator(),
-                )
-              else
-                _buildHitArea(),
+                ),
               Column(
                 mainAxisAlignment: MainAxisAlignment.end,
                 children: <Widget>[
@@ -129,7 +127,7 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
   @override
   void didChangeDependencies() {
     final oldController = _chewieController;
-    _chewieController = ChewieController.of(context);
+    _chewieController = ChewieAudioController.of(context);
     controller = chewieController.videoPlayerController;
 
     if (oldController != chewieController) {
@@ -183,7 +181,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
             await showModalBottomSheet<OptionItem>(
               context: context,
               isScrollControlled: true,
-              useRootNavigator: chewieController.useRootNavigator,
               builder: (context) => OptionsDialog(
                 options: options,
                 cancelButtonText:
@@ -192,9 +189,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
             );
           }
 
-          if (_latestValue.isPlaying) {
-            _startHideTimer();
-          }
         },
         icon: Icon(
           icon ?? Icons.more_vert,
@@ -274,7 +268,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
                       _buildSubtitleToggle(icon: Icons.subtitles),
                     if (chewieController.showOptions)
                       _buildOptionsButton(icon: Icons.settings),
-                    if (chewieController.allowFullScreen) _buildExpandButton(),
                   ],
                 ),
               ),
@@ -300,65 +293,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
     );
   }
 
-  GestureDetector _buildExpandButton() {
-    return GestureDetector(
-      onTap: _onExpandCollapse,
-      child: AnimatedOpacity(
-        opacity: notifier.hideStuff ? 0.0 : 1.0,
-        duration: const Duration(milliseconds: 300),
-        child: Container(
-          height: barHeight + (chewieController.isFullScreen ? 15.0 : 0),
-          margin: const EdgeInsets.only(right: 12.0),
-          padding: const EdgeInsets.only(
-            left: 8.0,
-            right: 8.0,
-          ),
-          child: Center(
-            child: Icon(
-              chewieController.isFullScreen
-                  ? Icons.fullscreen_exit
-                  : Icons.fullscreen,
-              color: Colors.white,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildHitArea() {
-    final bool isFinished = _latestValue.position >= _latestValue.duration;
-    final bool showPlayButton =
-        widget.showPlayButton && !_dragging && !notifier.hideStuff;
-
-    return GestureDetector(
-      onTap: () {
-        if (_latestValue.isPlaying) {
-          if (_displayTapped) {
-            setState(() {
-              notifier.hideStuff = true;
-            });
-          } else {
-            _cancelAndRestartTimer();
-          }
-        } else {
-          _playPause();
-
-          setState(() {
-            notifier.hideStuff = true;
-          });
-        }
-      },
-      child: CenterPlayButton(
-        backgroundColor: Colors.black54,
-        iconColor: Colors.white,
-        isFinished: isFinished,
-        isPlaying: controller.value.isPlaying,
-        show: showPlayButton,
-        onPressed: _playPause,
-      ),
-    );
-  }
 
   Future<void> _onSpeedButtonTap() async {
     _hideTimer?.cancel();
@@ -366,7 +300,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
     final chosenSpeed = await showModalBottomSheet<double>(
       context: context,
       isScrollControlled: true,
-      useRootNavigator: chewieController.useRootNavigator,
       builder: (context) => PlaybackSpeedDialog(
         speeds: chewieController.playbackSpeeds,
         selected: _latestValue.playbackSpeed,
@@ -375,10 +308,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
 
     if (chosenSpeed != null) {
       controller.setPlaybackSpeed(chosenSpeed);
-    }
-
-    if (_latestValue.isPlaying) {
-      _startHideTimer();
     }
   }
 
@@ -455,7 +384,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
 
   void _cancelAndRestartTimer() {
     _hideTimer?.cancel();
-    _startHideTimer();
 
     setState(() {
       notifier.hideStuff = false;
@@ -469,9 +397,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
 
     _updateState();
 
-    if (controller.value.isPlaying || chewieController.autoPlay) {
-      _startHideTimer();
-    }
 
     if (chewieController.showControlsOnInitialize) {
       _initTimer = Timer(const Duration(milliseconds: 200), () {
@@ -480,20 +405,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
         });
       });
     }
-  }
-
-  void _onExpandCollapse() {
-    setState(() {
-      notifier.hideStuff = true;
-
-      chewieController.toggleFullScreen();
-      _showAfterExpandCollapseTimer =
-          Timer(const Duration(milliseconds: 300), () {
-        setState(() {
-          _cancelAndRestartTimer();
-        });
-      });
-    });
   }
 
   void _playPause() {
@@ -518,17 +429,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
           controller.play();
         }
       }
-    });
-  }
-
-  void _startHideTimer() {
-    final hideControlsTimer = chewieController.hideControlsTimer.isNegative
-        ? ChewieController.defaultHideControlsTimer
-        : chewieController.hideControlsTimer;
-    _hideTimer = Timer(hideControlsTimer, () {
-      setState(() {
-        notifier.hideStuff = true;
-      });
     });
   }
 
@@ -583,7 +483,6 @@ class _MaterialDesktopControlsState extends State<MaterialDesktopControls>
             _dragging = false;
           });
 
-          _startHideTimer();
         },
         colors: chewieController.materialProgressColors ??
             ChewieProgressColors(